[WIP] migrating Primer::LabelComponent to Primer::Beta::Label (#1530)
