+++ .changeset/tasty-needles-deliver.md
@@ -1,0 +1,5 @@
+---
+'@primer/view-components': patch
+---
+
+migrating Primer::LabelComponent to Primer::Beta::Label
