+++ static/audited_at.json
@@ -47,6 +47,7 @@
   "Primer::Beta::Flash": "",
   "Primer::Beta::Heading": "",
   "Primer::Beta::IconButton": "",
+  "Primer::Beta::Label": "",
   "Primer::Beta::Text": "",
   "Primer::Beta::Truncate": "",
   "Primer::Beta::Truncate::TruncateText": "",
