+++ static/constants.json
@@ -574,6 +574,68 @@
       "invisible"
     ]
   },
+  "Primer::Beta::Label": {
+    "DEFAULT_SCHEME": "default",
+    "DEFAULT_SIZE": "medium",
+    "DEFAULT_TAG": "span",
+    "DEFAULT_VARIANT": "none",
+    "DEPRECATED_SCHEME_OPTIONS": [
+      "info",
+      "warning",
+      "orange",
+      "purple"
+    ],
+    "DEPRECATED_VARIANT_OPTIONS": [
+      "large",
+      "inline"
+    ],
+    "INLINE_CLASS": "Label--inline",
+    "SCHEME_MAPPINGS": {
+      "default": "",
+      "primary": "Label--primary",
+      "secondary": "Label--secondary",
+      "accent": "Label--accent",
+      "success": "Label--success",
+      "attention": "Label--attention",
+      "danger": "Label--danger",
+      "severe": "Label--severe",
+      "done": "Label--done",
+      "sponsors": "Label--sponsors",
+      "info": "Label--info",
+      "warning": "Label--warning",
+      "orange": "Label--orange",
+      "purple": "Label--purple"
+    },
+    "SCHEME_OPTIONS": [
+      "default",
+      "primary",
+      "secondary",
+      "accent",
+      "success",
+      "attention",
+      "danger",
+      "severe",
+      "done",
+      "sponsors"
+    ],
+    "SIZE_MAPPINGS": {
+      "medium": null,
+      "large": "Label--large"
+    },
+    "SIZE_OPTIONS": [
+      "medium",
+      "large"
+    ],
+    "TAG_OPTIONS": [
+      "span",
+      "summary",
+      "a",
+      "div"
+    ],
+    "VARIANT_OPTIONS": [
+      "none"
+    ]
+  },
   "Primer::Beta::Text": {
     "DEFAULT_TAG": "span"
   },
@@ -696,66 +758,6 @@
   "Primer::ImageCrop": {
   },
   "Primer::LabelComponent": {
-    "DEFAULT_SCHEME": "default",
-    "DEFAULT_SIZE": "medium",
-    "DEFAULT_TAG": "span",
-    "DEFAULT_VARIANT": "none",
-    "DEPRECATED_SCHEME_OPTIONS": [
-      "info",
-      "warning",
-      "orange",
-      "purple"
-    ],
-    "DEPRECATED_VARIANT_OPTIONS": [
-      "large",
-      "inline"
-    ],
-    "INLINE_CLASS": "Label--inline",
-    "SCHEME_MAPPINGS": {
-      "default": "",
-      "primary": "Label--primary",
-      "secondary": "Label--secondary",
-      "accent": "Label--accent",
-      "success": "Label--success",
-      "attention": "Label--attention",
-      "danger": "Label--danger",
-      "severe": "Label--severe",
-      "done": "Label--done",
-      "sponsors": "Label--sponsors",
-      "info": "Label--info",
-      "warning": "Label--warning",
-      "orange": "Label--orange",
-      "purple": "Label--purple"
-    },
-    "SCHEME_OPTIONS": [
-      "default",
-      "primary",
-      "secondary",
-      "accent",
-      "success",
-      "attention",
-      "danger",
-      "severe",
-      "done",
-      "sponsors"
-    ],
-    "SIZE_MAPPINGS": {
-      "medium": null,
-      "large": "Label--large"
-    },
-    "SIZE_OPTIONS": [
-      "medium",
-      "large"
-    ],
-    "TAG_OPTIONS": [
-      "span",
-      "summary",
-      "a",
-      "div"
-    ],
-    "VARIANT_OPTIONS": [
-      "none"
-    ]
   },
   "Primer::LayoutComponent": {
     "ALLOWED_SIDEBAR_COLS": [
