+++ static/statuses.json
@@ -47,6 +47,7 @@
   "Primer::Beta::Flash": "beta",
   "Primer::Beta::Heading": "beta",
   "Primer::Beta::IconButton": "beta",
+  "Primer::Beta::Label": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
   "Primer::Beta::Truncate::TruncateText": "alpha",
@@ -66,7 +67,7 @@
   "Primer::IconButton": "deprecated",
   "Primer::Image": "deprecated",
   "Primer::ImageCrop": "deprecated",
-  "Primer::LabelComponent": "beta",
+  "Primer::LabelComponent": "deprecated",
   "Primer::LayoutComponent": "alpha",
   "Primer::LinkComponent": "beta",
   "Primer::LocalTime": "alpha",
